Refactor names of files in the context folder and their imports
