+++ app/(pages)/(content)/administration/layout.tsx
@@ -2,7 +2,7 @@
 
 import { Roboto } from "next/font/google";
 import "@/styles/globals.css";
-import { AuthProvider, useAuth } from "@/contexts/authContext";
+import { AuthProvider, useAuth } from "@/contexts/auth-context";
 import NavbarAdmin from "@/components/navbar-admin";
 import Footer from "@/components/footer";
 import { useRouter } from "next/navigation";
