+++ app/(pages)/(content)/missing/page.tsx
@@ -8,10 +8,10 @@ import dynamic from 'next/dynamic';
 import Footer from "@/components/footer";
 import Image from "next/image";
 import { MapProps } from "@/types/map-props";
-import { useAppContext } from "@/contexts/appContext";
+import { useAppContext } from "@/contexts/app-context";
 //import Axios from "@/utils/axiosInstace";
 import { getAnimals } from "@/utils/animals.http";
-import { useAuth } from '@/contexts/authContext';
+import { useAuth } from '@/contexts/auth-context';
 import { getBreed } from "@/utils/breed.http";
 import { postPets } from "@/utils/pets.http";
 import { postMedia } from "@/utils/media.http";
