+++ app/(pages)/(content)/profile/[id]/page.tsx
@@ -10,7 +10,7 @@ import { UserProfile } from '@/types/user-profile';
 import { Post } from '@/types/post';
 import { Pet } from '@/types/pet';
 import { useParams, useRouter } from 'next/navigation';
-import { useAuth } from '@/contexts/authContext';
+import { useAuth } from '@/contexts/auth-context';
 import Loading from '@/app/loading';
 import { Detail } from '@/components/profile/detail-form';
 import { DropdownMenuButtons } from '@/components/profile/dropdown-buttons';
