+++ components/animal-breed-modal.tsx
@@ -2,7 +2,7 @@
 
 import { useState, useEffect } from "react";
 import { createBreed, updateBreed, deleteBreed } from "@/utils/breeds.http";
-import { useAuth } from "@/contexts/authContext";
+import { useAuth } from "@/contexts/auth-context";
 import Button from "@/components/buttons/button";
 import ConfirmationModal from "@/components/confirm-modal"; // Importa el modal de confirmación
