+++ components/post/post-content.tsx
@@ -1,7 +1,7 @@
 import PostLocationMap from "@/components/post/post-location-map";
 import PostComments from "@/components/post/post-comments";
 import { Post } from "@/types/post";
-import { useAuth } from "@/contexts/authContext";
+import { useAuth } from "@/contexts/auth-context";
 import { Pet } from "@/types/pet";
 
 interface PostContentProps {
